2023.04.25

+++ server/src/main/java/com/shikkeram/server/services/FurnitureService.java
@@ -5,28 +5,32 @@ import com.shikkeram.server.models.Furniture;
 import com.shikkeram.server.models.Image;
 import com.shikkeram.server.repositories.FurnitureCriteriaRepository;
 import com.shikkeram.server.repositories.FurnitureRepository;
-import com.shikkeram.server.repositories.ImagesRepository;
 import com.shikkeram.server.searchCriterias.FurniturePage;
 import com.shikkeram.server.searchCriterias.FurnitureSearchCriteria;
+import com.shikkeram.server.utils.Paths;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 @Transactional(readOnly = true)
 public class FurnitureService {
     private final FurnitureRepository repository;
-    private final ImagesRepository imagesRepository;
     private final FurnitureCriteriaRepository criteriaRepository;
+    private final FileService fileService;
 
-
-    public List<Furniture> findAll() {
-        return repository.findAll();
-    }
+    @Value("${serverDomain}")
+    private String serverDomain;
 
     public Page<Furniture> findWithFilters(FurniturePage page,
                                            FurnitureSearchCriteria furnitureSearchCriteria) {
@@ -40,30 +44,58 @@ public class FurnitureService {
     }
 
     @Transactional
-    public void save(Furniture furniture) {
-        List<Image> images = furniture.getImages();
-        repository.save(furniture);
+    public void save(Furniture furniture, List<MultipartFile> files) {
+        String[] imagePaths;
+        try {
+            imagePaths = fileService.save(files);
+        } catch (IOException e) {
+            throw new IllegalStateException();
+        }
 
-        images.forEach(i -> i.setFurniture(furniture));
-        imagesRepository.saveAll(images);
+        List<Image> images = new ArrayList<>(imagePaths.length);
+        for (int i = 0; i < imagePaths.length; i++) {
+            images.add(new Image(pathToLink(imagePaths[i]),i == 0, furniture));
+        }
+
+        furniture.setImages(images);
+
+        repository.save(furniture);
     }
 
     @Transactional
     public void delete(int id) {
-        throwNotFoundIfNotExists(id);
-        repository.deleteById(id);
-    }
+        Optional<Furniture> furniture = repository.findById(id);
+        if (furniture.isPresent()) {
+            List<String> paths = furniture.get().getImages()
+                    .stream()
+                    .map(path -> linkToPath(path.getLink()))
+                    .toList();
 
-    @Transactional
-    public void update(int id, Furniture updated) {
-        throwNotFoundIfNotExists(id);
+            fileService.deleteFiles(paths);
 
-        updated.setId(id);
-        repository.save(updated);
-    }
-
-    private void throwNotFoundIfNotExists(int id) {
-        if(repository.existsById(id))
+            furniture.get().removeImagesIfExists();
+            repository.deleteById(id);
+        } else {
             throw new EntityNotFoundException();
+        }
+    }
+
+//    @Transactional
+//    public void update(int id, Furniture updated) {
+//        throwNotFoundIfNotExists(id);
+//
+//        updated.setId(id);
+//        repository.save(updated);
+//    }
+
+    private String pathToLink(String path) {
+        return serverDomain +
+                "/images/" +
+                Path.of(path).getFileName();
+    }
+
+    private String linkToPath(String link) {
+        return Paths.UPLOAD_DIRECTORY +
+                link.substring(29);
     }
 }
